Add flexibility to launching parallel tests.
Not all mpirun/mpiexec implementations allow/use command-line arguments
to specify the number of processors to run on.  Allow this to be an
optional part of the parallel command template.  A consequence of this
is that testcode will have no control of the number of processors a job
is run on such systems.