Move old output files out of the way if required.
Previously we threw an error if the self.output was a globbing pattern
and matched several files.  This can easily occur if, for example,
a previous run is interrupted and several tests are contained within
a single directory.  To avoid this issue and potential data loss we now
move any files matching the globbing pattern to a backup directory.