Replace all spaces and / in testcode filenames with _.
Previously only spaces in the argument string were replaced and (more
importantly) slashes were unaltered, which caused problems as bash would
then interpret the filename to be a path including a subdirectory (which
doesn't exist) when trying to run the test.  This now allows input files
to reside in a different directory to where the test is running.