Enable nested tuples to be entered in the config files.
This is required for such values to be entered over multiple lines.
Previously we appended a comma to all affected elements to force
literal_eval to return a nested set of tuples, even if the config value
was only a single tuple.  However, this meant that a tuple of tuples
(the desired value!) could not be manually entered.  We now detect if
the supplied value is or isn't already a tuple of tuples and act
accordingly.