Force tests to run consecutively if required for safety.
If tests are run in the same directory with run_concurrent set to true
*and* multiple tests are run at the same time *and* the test program
writes its output to a specific file rather than STDOUT (so the output
variable is set) *and* wildcards are used in the value of the output
variable in the config files, then there's a race condition in running
tests and renaming the output files to the test output files.  As
a result, tests could fail to run or (worse) be given the wrong output
files to compare to benchmark.

For now, force affected tests to serialize (run consecutively).  To do:
add support for providing more information about the output filename so
wildcards aren't as necessary.

Reported by: Keith Refson.