skipped command added.
The skipped command is ran before extracting data from output files.  If
the skipped command returns 0, then the test is assumed to be irrelevant
(e.g. this version of the program does not have the features implemented
required for this test).

For example, if

skip_program = grep
skip_args = "is not implemented."

then testcode will run:

grep "is not implemented" test_file

where test_file is the test output file.  If grep returns 0 (i.e.
test_file contains the string "is not implemented") then the test is
marked as skipped and the test file is not compared to the benchmark.