Print more details for non-comparable tests.
If a test is not comparable to the benchmark (i.e. the data items in
the test and benchmark are different), we print out the name of the
fields that exists only in the test and not the benchmark or vice versa.
Then proceed to perform a standard validation of data items that
occur in both the benchmark and test.

These messages are printed out only if the verbose level is
greater than one.