Allow multiple benchmarks to be specified.
The first benchmark in the list in the userconfig file which exists for a given test is
used.  For now, the benchmark command-line option can still only be used to select
a single benchmark.  Need some experience to see if enabling that option to specify
alternate benchmark files would be useful.