Be more forgiving of executable locations under Windows/Cygwin.
os.path under Cygwin looks for unix-style paths but some MPI launchers (e.g. Intel)
expect a Windows-style path, which caused testcode to not find the binary, even
though the user gave it a path which would run correctly.

I don't have access to a Windows machine and am reluctant to add lots of
cygwin-specifc checking.  Deciding if a Windows-style or Unix-style path is
acceptable seems painful and fragile.  Instead, be less cautious under
Windows/Cygwin and attempt to run the calculations even if the binary cannot be
found.

Reported by: Keith Refson