Detect if a data set has different amounts of data in test/benchmark.
We were detecting if the test or output contains a different number of
keys labelling the data, but not if an individual key contained
a different amount of data (i.e. if data item 'AAA' contains data
[1,2,3] in the benchmark but [1,2,3,4] in the test).  Now test for this explicitly.

Note that this could, in corner cases, have led to tests passing when
they should have failed.

Reported by: José María Escartín Esteban.