Exclude ignore_fields in checking for the same amount of data.
The change to check that the test and benchmark outputs contained the
same amount of data failed to take into account the optional fields to
be ignored, causing an exception to be raised.  Handle this by making
sure that the test and data outputs contain the same amount of data for
all other fields bar those specified in ignore_fields.

Introduced in 41e016ac.