Fill in any missing data columns when printing data tables.
We only combine data tables if they contain the same data to be tested.
However, if a column is missing but is not included in the tests, this
led to an exception when we attempted to print out data not present in
one or other of the test and benchmark.

Instead, in such cases, fill in the data with 'n/a' entries.

Reported by: Nick Blunt.