Select tests from the unique list of test categories.
Previously we were using the entire list of categories defined in
jobconfig.  For some cases (in particular, those making heavy use of the
run_concurrent option) which contain multiple entries for the same
category, this leads to a large number (N^2) costly comparisons.  We can
large avoid this by only checking the set, rather than list, of all
categories.

This change reduces the time taken by

$ testcode.py compare -q -c simple

for CASTEP 8.0 from 140s to 28s.