Ability to extract data from an arbitrary python function.
Using the importlib functionality in python 2.7/3.1+, we can import
a user-specified module and hence run a function to extract the data
rather than going via subprocess.

If the data analysis code requires importing a large module (e.g.
pandas) then this can be *substantially* faster, as the repeated imports
are avoided.